Move all public client functions to api module #148
